Fixed ScanAll() result slice capacity.
The previous code was not buggy, but redundant
and ineffective. The size of the array underlying
the result slice was set to 0 because this is
the start of the scanner cursor right after it's
initialized. Now the capacity matches the number
of bytes in the buffer, which should be optimal
for most cases given that multi-byte unicode
runes are less likely to occur.